requirements.txt
python-binance==1.0.19
requests
flask
matplotlib
numpy
websocket-client